+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Enhancements
 * Remove type coercion on bool and ObjectId when doing queries.
+* Pass CreatePolicy to `unbox<T>` from the object accessor.
 
 ### Fixed
 * <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-core/issues/????), since v?.?.?)